Add basic unit testing
e91bb69bfd071f5ea9422e876fd8538100ab8a01 introduced a bug where we stopped
reading the fontlock names from our data files correctly. This adds a test for
this case, but more importantly it adds a unit testing framework to the whole
package. This won't get picked up for distribution, but is used for verifying
the correctness of the code in CI.